// front/src/app/core/auth/models/application-paths-type.model.ts
export interface ApplicationPathsType {
    readonly DefaultLoginRedirectPath: string;
    readonly ApiAuthorizationClientConfigurationUrl: string;
    readonly Login: string;
    readonly LoginFailed: string;
    readonly LoginCallback: string;
    readonly Register: string;
    readonly Profile: string;
    readonly LogOut: string;
    readonly LoggedOut: string;
    readonly LogOutCallback: string;
    readonly LoginPathComponents: string [];
    readonly LoginFailedPathComponents: string [];
    readonly LoginCallbackPathComponents: string [];
    readonly RegisterPathComponents: string [];
    readonly ProfilePathComponents: string [];
    readonly LogOutPathComponents: string [];
    readonly LoggedOutPathComponents: string [];
    readonly LogOutCallbackPathComponents: string [];
    readonly IdentityRegisterPath: string;
    readonly IdentityManagePath: string;
  }
  